Added automatic PREFIX detection
Added automatic PREFIX detection so the user will not have to input the PREFIX when they want to use the script. PREFIX can still be set manually because due to the nature of the code, in a scene that might, say, have 2 Mineways imports, it will pick the one with the most diverse amounts of blocks.

Also clarified a print statement.